Closes #33
Wrote new hold spawner script and made a new version of sampler script.  The hold spawner spawns a normal arrow at the top on enable then spawns hold arrows behind it continuously on a given spawn frequency based on framerate until disable is called again.  The new hold detecting spawner script keeps the hold spawns active until another black space is detected on the video it is sampling from and then disables it to work with the hold spawner script.