Closes #28 arrow clicking functionality
Added colliders and a script that detects entry of arrows and deletes them as well as records the distance apart they are for use in scoring later.  Added to the script that lights up stars on clicks to also activate these collider objects with the new scripts.  Changed tagging on arrow prefabs